+++ Tests/NukeTests/ImagePipelineTests/ImagePipelineLoadDataTests.swift
@@ -42,11 +42,7 @@ struct ImagePipelineLoadDataTests {
         )
 
         // When
-        let task = pipeline.imageTask(with: Test.url)
-        var progressValues: [ImageTask.Progress] = []
-        for await progress in task.progress {
-            progressValues.append(progress)
-        }
+        let (task, progressValues) = await pipeline.recordProgress(for: ImageRequest(url: Test.url))
         _ = try? await task.response
 
         // Then